os2: open files explicitly in binary mode
-Zbin-files linker flag is a very convenient way. If this flag is used
to the executables such as .exe and .dll, it's fine. However, in case of
a static lib, it has no effect. As a result, other programs linked
aginst a static lib are forced to use -Zbin-files linker flag, too.
Unless it is specified, the result is unexpectable.

After all, it's the better way to open files explicitly in binary mode
than to use -Zbin-files linker flag.

    modified:   configure.ac
    modified:   libarchive/archive_util.c